@@ -11,11 +11,15 @@ export const actions = {
 	REQUEST_DATA: 'REQUEST_DATA',
 	DATA_RECEIVED: 'DATA_RECEIVED',
 	CLEAR_DATA: 'CLEAR_DATA',
-	FILTER_LIST: 'FILTER_LIST',
+	SET_CITY_FILTER_LIST: 'SET_CITY_FILTER_LIST',
+	SET_DISTANCE_FILTER: 'SET_DISTANCE_FILTER',
 	RESET_FILTER: 'RESET_FILTER',
 	LOGOUT: 'LOGOUT',
+	SET_USERNAME: 'SET_USERNAME',
+	SET_PASSWORD: 'SET_PASSWORD',
 };
 
+export const checkLocalStorage = createAction(actions.CHECK_LOCAL_STORAGE);
 export const loginAction = createAction(actions.LOGIN);
 export const loginSeccess = createAction(actions.LOGIN_SUCCESS);
 export const requestData = createAction(actions.REQUEST_DATA);
@@ -23,16 +27,10 @@ export const clearData = createAction(actions.CLEAR_DATA);
 export const dataReceived = createAction(actions.DATA_RECEIVED);
 export const resetFilter = createAction(actions.RESET_FILTER);
 export const logout = createAction(actions.LOGOUT);
-
-export const filterList = (city, distance) => (
-	{
-		type: actions.FILTER_LIST,
-		payload: {
-			city,
-			distance
-		}
-	}
-);
+export const setCityFilter = createAction(actions.SET_CITY_FILTER_LIST);
+export const setDistanceFilter = createAction(actions.SET_DISTANCE_FILTER);
+export const setUsername = createAction(actions.SET_USERNAME);
+export const setPassword = createAction(actions.SET_PASSWORD);
 
 export const login = (username, password) => {
 	return function(dispatch) {
