@@ -1,9 +1,12 @@
 import { connect } from 'react-redux';
 
+import { getAuthorizationState, getUsername, getPassword } from '../../selectors';
 import App from './App';
 
 const mapStateToProps = state => ({
-	isAuthorized: state.authorized.isAuthorized
+	username: getUsername(state),
+	password: getPassword(state),
+	isAuthorized: getAuthorizationState(state)
 });
 
 const AppContainer = connect(mapStateToProps)(App);
