@@ -6,45 +6,34 @@ import Button from '../Button/Button';
 
 import './Filter.scss';
 
-const Filter = ({ onCityFilterInput, onDistanceFilterInput, resetFilter }) => {
-	let cityInput = React.createRef();
-	let distanceInput = React.createRef();
-
-	function resetInputFields() {
-		cityInput.current.clearInput();
-		distanceInput.current.clearInput();
-	}
-
-	return (
-		<section className="dashboard__filter">
-			<div className="dashboard__filter__input-container">
-				<Input
-					subclass={'filter-city'}
-					type={'text'}
-					placeholder={'City'}
-					inputHandler={onCityFilterInput}
-					ref={cityInput}
-				/>
-				<Input
-					subclass={'filter-distance'}
-					type={'number'}
-					placeholder={'Distance'}
-					inputHandler={onDistanceFilterInput}
-					ref={distanceInput}
-				/>
-			</div>
-			<Button
-				text={'Reset'}
-				onClick={() => {
-					resetInputFields();
-					resetFilter();
-				}}
+const Filter = ({ city, distance, onCityFilterInput, onDistanceFilterInput, resetFilter }) => (
+	<section className="dashboard__filter">
+		<div className="dashboard__filter__input-container">
+			<Input
+				subclass={'filter-city'}
+				type={'text'}
+				placeholder={'City'}
+				inputHandler={onCityFilterInput}
+				value={city}
 			/>
-		</section>
-	);
-};
+			<Input
+				subclass={'filter-distance'}
+				type={'number'}
+				placeholder={'Distance'}
+				inputHandler={onDistanceFilterInput}
+				value={distance}
+			/>
+		</div>
+		<Button
+			text={'Reset'}
+			onClick={() => resetFilter()}
+		/>
+	</section>
+);
 
 Filter.propTypes = {
+	city: PropTypes.string,
+	distance: PropTypes.number,
 	resetFilter: PropTypes.func,
 	onCityFilterInput: PropTypes.func,
 	onDistanceFilterInput: PropTypes.func,
